Allow specifying a file to be parsed to the CLI
