+++ cli/Cargo.toml
@@ -11,3 +11,4 @@ parser = { path = "../parser" }
 miette = { version = "5.5.0", features = ["fancy"] }
 thiserror = "1.0.38"
 dbg-pls = { version = "0.3.5", features = ["pretty", "colors"] }
+clap = { version = "4.2.2", features = ["derive"] }
