+++ cli/src/cli.rs
@@ -1,0 +1,11 @@
+use std::path::PathBuf;
+
+use clap::Parser;
+
+#[derive(Parser)]
+#[command(author, version, about, long_about = None)]
+pub struct Cli {
+    /// Defines the source file to parse
+    #[arg(short, long, value_name = "FILE")]
+    pub(crate) source: Option<PathBuf>,
+}
